# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for VideoCall app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateRoomBody {
  /** Optional display name for the host */
  hostName?: string;
}

export interface Room {
  id: string;
  hostName?: string | null;
  participantCount: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}