# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for VideoCall app
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new video call room
 */
export const CreateRoomBody = zod.object({
  hostName: zod
    .string()
    .optional()
    .describe("Optional display name for the host"),
});

/**
 * @summary Get room info
 */
export const GetRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.string(),
  hostName: zod.string().nullish(),
  participantCount: zod.number(),
  createdAt: zod.coerce.date(),
});