# lib/api-zod/src/generated/types/createRoomBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for VideoCall app
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomBody {
  /** Optional display name for the host */
  hostName?: string;
}